refactor: set v8 flags before initialize v8 platform
After the v8 platform had initialized, it was no longer able to set the
v8 flags, so it is necessary to adjust the invocation order of the set
v8 flag function to avoid v8's static assertion.

(cherry picked from commit bd1975276df6f3ded90a23c21e5729ebe50a2dba)
