+++ crates/base/src/deno_runtime.rs
@@ -58,20 +58,6 @@ fn get_error_class_name(e: &AnyError) -> &'static str {
     sb_core::errors_rt::get_error_class_name(e).unwrap_or("Error")
 }
 
-fn set_v8_flags() {
-    let v8_flags = std::env::var("V8_FLAGS").unwrap_or("".to_string());
-    let mut vec = vec!["IGNORED"];
-    if v8_flags.is_empty() {
-        return;
-    }
-
-    vec.append(&mut v8_flags.split(' ').collect());
-    error!(
-        "v8 flags unrecognized {:?}",
-        deno_core::v8_set_flags(vec.iter().map(|v| v.to_string()).collect())
-    );
-}
-
 pub struct DenoRuntime {
     pub js_runtime: JsRuntime,
     pub env_vars: HashMap<String, String>, // TODO: does this need to be pub?
@@ -94,8 +80,6 @@ impl DenoRuntime {
             maybe_entrypoint,
             maybe_module_code,
         } = opts;
-
-        set_v8_flags();
 
         let user_agent = "supabase-edge-runtime".to_string();
         let base_dir_path = std::env::current_dir().map(|p| p.join(&service_path))?;
